+++ tdms/include/output_matrices.h
@@ -34,8 +34,6 @@ private:
    * @param matrix_names List of output matrices to check.
    */
   bool memory_already_assigned(std::vector<std::string> matrix_names);
-  /*! @copydoc memory_already_assigned */
-  bool memory_already_assigned(std::string matrix_name);
   /**
    * @brief Throw an error if the matrix_pointer corresponding to any one of the matrix names passed already points to allocated memory.
    *
@@ -48,12 +46,7 @@ private:
       throw std::runtime_error("Reassigning output pointer will cause memory leak - aborting.");
     }
   }
-  /*! @copydoc error_on_memory_assigned */
-  void error_on_memory_assigned(std::string matrix_name) {
-    if (memory_already_assigned(matrix_name)) {
-      throw std::runtime_error("Reassigning output pointer will cause memory leak - aborting.");
-    }
-  }
+
   /**
    * @brief Assign the output matrices provided as empty arrays.
    *
@@ -88,26 +81,6 @@ public:
     return matrix_pointers[index_from_matrix_name(matrix_name)];
   }
 
-  /**
-   * @brief Set the matrix pointer with the given index.
-   *
-   * Not recommended unless you're certain you are setting the correct array pointer! Use the overload which takes the array name as an input if you are uncertain you are setting the correct pointer.
-   *
-   * This function mainly sees use during destruction (and output writing) when we just want to iterate over all the matrices that we are dealing with.
-   *
-   * @param index The index in matrix_pointers to set
-   * @param new_ptr The address we want to assign
-   */
-  void set_matrix_pointer(int index, mxArray *new_ptr) { matrix_pointers[index] = new_ptr; }
-  /**
-   * @brief Set the pointer to the named matrix
-   *
-   * @param matrix_name The name of the matrix that we want to set the pointer to
-   * @param new_ptr The address we want to assign
-   */
-  void set_matrix_pointer(const std::string &matrix_name, mxArray *new_ptr) {
-    set_matrix_pointer(index_from_matrix_name(matrix_name), new_ptr);
-  }
   /**
    * @brief Set the maxresfield object. If memory has not been reserved yet, it can be assigned here.
    *
